refactor tracing across multiple files
