+++ packages/chatbot-server-mongodb-public/src/tracing/routesUpdateTraceHandlers.ts
@@ -1,0 +1,79 @@
+import { strict as assert } from "assert";
+import { UpdateTraceFunc } from "mongodb-chatbot-server/build/routes/conversations/UpdateTraceFunc";
+import { ObjectId } from "mongodb-rag-core/mongodb";
+import { extractTracingData } from "./extractTracingData";
+import { LlmAsAJudge, getLlmAsAJudgeScores } from "./getLlmAsAJudgeScores";
+
+export const makeAddMessageToConversationUpdateTrace: (
+  k: number,
+  llmAsAJudge?: LlmAsAJudge & {
+    /**
+    Percent of numbers to judge. Between 0 and 1.
+   */
+    percentToJudge: number;
+  }
+) => UpdateTraceFunc = (k, llmAsAJudge) => {
+  validatePercentToJudge(llmAsAJudge?.percentToJudge);
+
+  return async function ({ traceId, conversation, logger }) {
+    const tracingData = extractTracingData(
+      conversation.messages,
+      ObjectId.createFromHexString(traceId)
+    );
+    const shouldJudge =
+      typeof llmAsAJudge?.percentToJudge === "number" &&
+      Math.random() < llmAsAJudge.percentToJudge;
+
+    logger.updateSpan({
+      id: traceId,
+      tags: tracingData.tags,
+      scores: {
+        ...getTracingScores(tracingData, k),
+        ...(shouldJudge
+          ? await getLlmAsAJudgeScores(llmAsAJudge, tracingData)
+          : undefined),
+      },
+    });
+  };
+};
+
+function getTracingScores(
+  tracingData: ReturnType<typeof extractTracingData>,
+  k: number
+) {
+  return {
+    RejectedQuery: tracingData.rejectQuery === true ? 1 : null,
+    VerifiedAnswer: tracingData.isVerifiedAnswer === true ? 1 : null,
+    LlmDoesNotKnow: tracingData.llmDoesNotKnow === true ? 1 : null,
+    [`RetrievedChunksOver${k}`]:
+      tracingData.isVerifiedAnswer !== true
+        ? tracingData.numRetrievedChunks / k
+        : null,
+  };
+}
+
+export function makeCommentMessageUpdateTrace(
+  llmAsAJudge: LlmAsAJudge
+): UpdateTraceFunc {
+  return async function ({ traceId, conversation, logger }) {
+    logger.updateSpan({
+      id: traceId,
+      scores: await getLlmAsAJudgeScores(
+        llmAsAJudge,
+        extractTracingData(
+          conversation.messages,
+          ObjectId.createFromHexString(traceId)
+        )
+      ),
+    });
+  };
+}
+
+function validatePercentToJudge(percentToJudge?: number) {
+  if (percentToJudge !== undefined) {
+    assert(
+      percentToJudge >= 0 && percentToJudge <= 1,
+      `percentToJudge must be between 0 and 1. Received: ${percentToJudge}`
+    );
+  }
+}
